Disable snapshotting for Redis server
Fixes the following error:

MISCONF Redis is configured to save RDB snapshots, but it is currently
not able to persist on disk. Commands that may modify the data set are
disabled, because this instance is configured to report errors during
writes if RDB snapshotting fails (stop-writes-on-bgsave-error option).
Please check the Redis logs for details about the RDB error.

https://3.basecamp.com/3923521/buckets/6162994/todos/1140769347