TASK: Allow overwriting the hz configuration setting with env
The hz configuration determines how often redis operates background
processes like key reclaims and active expire checks.

In busy instances with many changes we might want to set this higher
than the default (10). This chnge will have no effect if the
environment variable is not set as we keep the default as is.